+++ library/src/commonMain/kotlin/org/jraf/klibnotion/client/NotionClient.kt
@@ -26,14 +26,14 @@ package org.jraf.klibnotion.client
 
 import org.jraf.klibnotion.internal.client.NotionClientImpl
 import org.jraf.klibnotion.model.base.UuidString
+import org.jraf.klibnotion.model.block.value.BlockValueList
+import org.jraf.klibnotion.model.block.value.BlockValueListProducer
 import org.jraf.klibnotion.model.database.Database
 import org.jraf.klibnotion.model.database.query.DatabaseQuery
 import org.jraf.klibnotion.model.database.query.DatabaseQuerySort
 import org.jraf.klibnotion.model.page.Page
 import org.jraf.klibnotion.model.pagination.Pagination
 import org.jraf.klibnotion.model.pagination.ResultPage
-import org.jraf.klibnotion.model.property.content.ContentValueList
-import org.jraf.klibnotion.model.property.content.ContentValueListProducer
 import org.jraf.klibnotion.model.property.value.PropertyValueList
 import org.jraf.klibnotion.model.user.User
 import kotlin.jvm.JvmStatic
@@ -108,7 +108,7 @@ interface NotionClient {
         suspend fun createPage(
             parentDatabaseId: UuidString,
             properties: PropertyValueList = PropertyValueList(),
-            content: ContentValueList? = null,
+            content: BlockValueList? = null,
         ): Page
 
         /**
@@ -118,7 +118,7 @@ interface NotionClient {
         suspend fun createPage(
             parentDatabaseId: UuidString,
             properties: PropertyValueList = PropertyValueList(),
-            content: ContentValueListProducer,
+            content: BlockValueListProducer,
         ): Page
 
         /**