+++ library/src/commonMain/kotlin/org/jraf/klibnotion/internal/api/model/page/ApiCreateTableParametersConverter.kt
@@ -28,25 +28,25 @@ import org.jraf.klibnotion.internal.api.model.ApiConverter
 import org.jraf.klibnotion.internal.api.model.content.value.ApiOutContentValueConverter
 import org.jraf.klibnotion.internal.api.model.modelToApi
 import org.jraf.klibnotion.internal.api.model.property.value.ApiOutPropertyValueConverter
-import org.jraf.klibnotion.model.property.content.ContentValue
+import org.jraf.klibnotion.model.block.value.BlockValue
 import org.jraf.klibnotion.model.property.value.PropertyValue
 
 internal object ApiCreateTableParametersConverter :
     ApiConverter<
             ApiCreateTableParameters,
-            Triple<String, List<PropertyValue<*>>, List<ContentValue>?>
+            Triple<String, List<PropertyValue<*>>, List<BlockValue>?>
             >() {
-    override fun modelToApi(model: Triple<String, List<PropertyValue<*>>, List<ContentValue>?>): ApiCreateTableParameters {
+    override fun modelToApi(model: Triple<String, List<PropertyValue<*>>, List<BlockValue>?>): ApiCreateTableParameters {
         val parentDatabaseId = model.first
         val properties = model.second
-        val content = model.third
+        val children = model.third
 
         return ApiCreateTableParameters(
             parent = ApiPageParent(
                 database_id = parentDatabaseId,
             ),
             properties = properties.associate(ApiOutPropertyValueConverter::modelToApi),
-            content = content?.modelToApi(ApiOutContentValueConverter),
+            children = children?.modelToApi(ApiOutContentValueConverter),
         )
     }
 }