+++ library/src/commonMain/kotlin/org/jraf/klibnotion/model/block/value/NumberedListItemBlockValue.kt
@@ -22,11 +22,9 @@
  * limitations under the License.
  */
 
-package org.jraf.klibnotion.model.property.content
+package org.jraf.klibnotion.model.block.value
 
 /**
  * See [https://www.notion.so/notiondevs/fa3660a1844b451aa99e9aac965438c1?v=9b36837a440f448cbd2dd39f12edcfba].
  */
-interface ToDoContentValue : ContentValue {
-    val checked: Boolean
-}
+interface NumberedListItemBlockValue : BlockValue