+++ library/src/commonMain/kotlin/org/jraf/klibnotion/model/block/value/ToggleBlockValue.kt
@@ -22,13 +22,9 @@
  * limitations under the License.
  */
 
-package org.jraf.klibnotion.internal.model.content.value
+package org.jraf.klibnotion.model.block.value
 
-import org.jraf.klibnotion.model.property.content.ContentValueList
-import org.jraf.klibnotion.model.property.content.NumberedListItemContentValue
-import org.jraf.klibnotion.model.richtext.RichTextList
-
-internal data class NumberedListItemContentValueImpl(
-    override val text: RichTextList,
-    override val content: ContentValueList?,
-) : NumberedListItemContentValue
+/**
+ * See [https://www.notion.so/notiondevs/fa3660a1844b451aa99e9aac965438c1?v=9b36837a440f448cbd2dd39f12edcfba].
+ */
+interface ToggleBlockValue : BlockValue