+++ library/src/jvmMain/kotlin/org/jraf/klibnotion/client/future/FutureNotionClient.kt
@@ -29,14 +29,14 @@ package org.jraf.klibnotion.client.future
 import org.jraf.klibnotion.client.NotionClient
 import org.jraf.klibnotion.internal.client.future.FutureNotionClientImpl
 import org.jraf.klibnotion.model.base.UuidString
+import org.jraf.klibnotion.model.block.value.BlockValueList
+import org.jraf.klibnotion.model.block.value.BlockValueListProducer
 import org.jraf.klibnotion.model.database.Database
 import org.jraf.klibnotion.model.database.query.DatabaseQuery
 import org.jraf.klibnotion.model.database.query.DatabaseQuerySort
 import org.jraf.klibnotion.model.page.Page
 import org.jraf.klibnotion.model.pagination.Pagination
 import org.jraf.klibnotion.model.pagination.ResultPage
-import org.jraf.klibnotion.model.property.content.ContentValueList
-import org.jraf.klibnotion.model.property.content.ContentValueListProducer
 import org.jraf.klibnotion.model.property.value.PropertyValueList
 import org.jraf.klibnotion.model.user.User
 import java.util.concurrent.Future
@@ -99,7 +99,7 @@ interface FutureNotionClient {
         fun createPage(
             parentDatabaseId: UuidString,
             properties: PropertyValueList = PropertyValueList(),
-            content: ContentValueList? = null,
+            content: BlockValueList? = null,
         ): Future<Page>
 
         /**
@@ -108,7 +108,7 @@ interface FutureNotionClient {
         fun createPage(
             parentDatabaseId: UuidString,
             properties: PropertyValueList = PropertyValueList(),
-            content: ContentValueListProducer,
+            content: BlockValueListProducer,
         ): Future<Page>
 
         /**