+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next (????-??-??)
 
 - Remove `isArchived` parameter from `getPage` following the Notion API change.
+- Fix write content API following the Notion API change.
 
 ## v1.3.0 (2021-04-03)